changed: latency value for sound devices to microseconds to match the rest of the timing, also it's now optional, without it the default portaudio latency is used
git-svn-id: http://boblight.googlecode.com/svn/trunk@257 c8956a1e-8086-11de-bc80-f7d40095c7a4

+++ src/device/devicesound.h
@@ -29,8 +29,8 @@ class CDeviceSound : public CDevice
   public:
     CDeviceSound(CClientsHandler& clients);
 
-    void SetLatency(float latency) { m_latency = latency; }
-    void SetPeriod(int period)     { m_period = period; }
+    void SetLatency(double latency) { m_latency = latency; }
+    void SetPeriod(int period)      { m_period = period; }
  
   private:
     bool SetupDevice();
